+++ ledger/tracker.go
@@ -81,13 +81,13 @@ type ledgerTracker interface {
 	// and the tracker is expected to still function after a
 	// restart and a call to loadFromDisk().  For example,
 	// returning 0 means that no blocks can be deleted.
-	committedUpTo(basics.Round) basics.Round
+	committedUpTo(basics.Round) (minRound, lookback basics.Round)
 
 	// prepareCommit, commitRound and postCommit are called when it is time to commit tracker's data.
 	// If an error returned the process is aborted.
 	prepareCommit(*deferredCommitContext) error
 	commitRound(context.Context, *sql.Tx, *deferredCommitContext) error
-	postCommit(deferredCommitContext)
+	postCommit(context.Context, *deferredCommitContext)
 
 	// handleUnorderedCommit is a special method for handling deferred commits that are out of order.
 	// Tracker might update own state in this case. For example, account updates tracker cancels
@@ -99,6 +99,9 @@ type ledgerTracker interface {
 	// be called even if loadFromDisk() is not called or does
 	// not succeed.
 	close()
+
+	// todo - document me
+	produceCommittingTask(committedRound basics.Round, dbRound basics.Round, dcc *deferredCommitContext) *deferredCommitContext
 }
 
 // ledgerForTracker defines the part of the ledger that a tracker can
@@ -115,10 +118,6 @@ type ledgerForTracker interface {
 	GenesisHash() crypto.Digest
 	GenesisProto() config.ConsensusParams
 	GenesisAccounts() map[basics.Address]basics.AccountData
-
-	// TODO: temporary?
-	scheduleCommit(basics.Round)
-	waitAccountsWriting()
 }
 
 type trackerRegistry struct {
@@ -129,8 +128,10 @@ type trackerRegistry struct {
 	ctx context.Context
 	// ctxCancel is the canceling function for canceling the committing go-routine ( i.e. signaling the committing go-routine that it's time to abort )
 	ctxCancel context.CancelFunc
-	// committedOffset is the offset at which we'd like to persist all the previous account information to disk.
-	committedOffset chan deferredCommit
+
+	// deferredCommits is the channel of pending deferred commits
+	deferredCommits chan *deferredCommitContext
+
 	// commitSyncerClosed is the blocking channel for synchronizing closing the commitSyncer goroutine. Once it's closed, the
 	// commitSyncer can be assumed to have aborted.
 	commitSyncerClosed chan struct{}
@@ -144,6 +145,12 @@ type trackerRegistry struct {
 
 	dbs db.Pair
 	log logging.Logger
+
+	// the synchronous mode that would be used for the account database.
+	synchronousMode db.SynchronousMode
+
+	// the synchronous mode that would be used while the accounts database is being rebuilt.
+	accountsRebuildSynchronousMode db.SynchronousMode
 
 	mu deadlock.RWMutex
 }
@@ -164,15 +171,19 @@ type deferredCommitContext struct {
 
 	updatedPersistedAccounts []persistedAccountData
 
-	isCatchpointRound    bool
-	committedRoundDigest crypto.Digest
-	trieBalancesHash     crypto.Digest
+	isCatchpointRound        bool
+	committedRoundDigest     crypto.Digest
+	trieBalancesHash         crypto.Digest
+	updatingBalancesDuration time.Duration
+	catchpointLabel          string
 
 	stats       telemetryspec.AccountsUpdateMetrics
 	updateStats bool
+
+	catchpointWriting *int32 // todo : doc.
 }
 
-func (tr *trackerRegistry) initialize(au *accountUpdates, l ledgerForTracker, trackers []ledgerTracker) (err error) {
+func (tr *trackerRegistry) initialize(au *accountUpdates, l ledgerForTracker, trackers []ledgerTracker, cfg config.Local) (err error) {
 	tr.driver = au
 
 	tr.dbs = l.trackerDB()
@@ -188,9 +199,11 @@ func (tr *trackerRegistry) initialize(au *accountUpdates, l ledgerForTracker, tr
 	}
 
 	tr.ctx, tr.ctxCancel = context.WithCancel(context.Background())
-	tr.committedOffset = make(chan deferredCommit, 1)
+	tr.deferredCommits = make(chan *deferredCommitContext, 1)
 	tr.commitSyncerClosed = make(chan struct{})
-	go tr.commitSyncer(tr.committedOffset)
+	tr.synchronousMode = db.SynchronousMode(cfg.LedgerSynchronousMode)
+	tr.accountsRebuildSynchronousMode = db.SynchronousMode(cfg.AccountsRebuildSynchronousMode)
+	go tr.commitSyncer(tr.deferredCommits)
 
 	tr.trackers = append(tr.trackers, trackers...)
 	return
@@ -210,7 +223,168 @@ func (tr *trackerRegistry) loadFromDisk(l ledgerForTracker) error {
 		}
 	}
 
-	return nil
+	err := tr.initializeTrackerCaches(l)
+	if err != nil {
+		return err
+	}
+	// the votes have a special dependency on the account updates, so we need to initialize these separetly.
+	err = tr.driver.voters.loadFromDisk(l, tr.driver)
+	return err
+}
+
+// initializeCaches fills up the accountUpdates cache with the most recent ~320 blocks ( on normal execution ).
+// the method also support balances recovery in cases where the difference between the lastBalancesRound and the lastestBlockRound
+// is far greater than 320; in these cases, it would flush to disk periodically in order to avoid high memory consumption.
+func (tr *trackerRegistry) initializeTrackerCaches(l ledgerForTracker) (err error) {
+	lastestBlockRound := l.Latest()
+	lastBalancesRound := tr.dbRound
+
+	var blk bookkeeping.Block
+	var delta ledgercore.StateDelta
+
+	accLedgerEval := accountUpdatesLedgerEvaluator{
+		au: tr.driver,
+	}
+	if lastBalancesRound < lastestBlockRound {
+		accLedgerEval.prevHeader, err = l.BlockHdr(lastBalancesRound)
+		if err != nil {
+			return err
+		}
+	}
+
+	skipAccountCacheMessage := make(chan struct{})
+	writeAccountCacheMessageCompleted := make(chan struct{})
+	defer func() {
+		close(skipAccountCacheMessage)
+		select {
+		case <-writeAccountCacheMessageCompleted:
+			if err == nil {
+				tr.log.Infof("initializeCaches completed initializing account data caches")
+			}
+		default:
+		}
+	}()
+
+	// this goroutine logs a message once if the parent function have not completed in initializingAccountCachesMessageTimeout seconds.
+	// the message is important, since we're blocking on the ledger block database here, and we want to make sure that we log a message
+	// within the above timeout.
+	go func() {
+		select {
+		case <-time.After(initializingAccountCachesMessageTimeout):
+			tr.log.Infof("initializeCaches is initializing account data caches")
+			close(writeAccountCacheMessageCompleted)
+		case <-skipAccountCacheMessage:
+		}
+	}()
+
+	blocksStream := make(chan bookkeeping.Block, initializeCachesReadaheadBlocksStream)
+	blockEvalFailed := make(chan struct{}, 1)
+	var blockRetrievalError error
+	go func() {
+		defer close(blocksStream)
+		for roundNumber := lastBalancesRound + 1; roundNumber <= lastestBlockRound; roundNumber++ {
+			blk, blockRetrievalError = l.Block(roundNumber)
+			if blockRetrievalError != nil {
+				return
+			}
+			select {
+			case blocksStream <- blk:
+			case <-blockEvalFailed:
+				return
+			}
+		}
+	}()
+
+	lastFlushedRound := lastBalancesRound
+	const accountsCacheLoadingMessageInterval = 5 * time.Second
+	lastProgressMessage := time.Now().Add(-accountsCacheLoadingMessageInterval / 2)
+
+	// rollbackSynchronousMode ensures that we switch to "fast writing mode" when we start flushing out rounds to disk, and that
+	// we exit this mode when we're done.
+	rollbackSynchronousMode := false
+	defer func() {
+		if rollbackSynchronousMode {
+			// restore default synchronous mode
+			tr.dbs.Wdb.SetSynchronousMode(context.Background(), tr.synchronousMode, tr.synchronousMode >= db.SynchronousModeFull)
+		}
+	}()
+
+	for blk := range blocksStream {
+		delta, err = l.trackerEvalVerified(blk, &accLedgerEval)
+		if err != nil {
+			close(blockEvalFailed)
+			return
+		}
+		tr.newBlock(blk, delta)
+
+		// flush to disk if any of the following applies:
+		// 1. if we have loaded up more than initializeCachesRoundFlushInterval rounds since the last time we flushed the data to disk
+		// 2. if we completed the loading and we loaded up more than 320 rounds.
+		flushIntervalExceed := blk.Round()-lastFlushedRound > initializeCachesRoundFlushInterval
+		loadCompleted := (lastestBlockRound == blk.Round() && lastBalancesRound+basics.Round(blk.ConsensusProtocol().MaxBalLookback) < lastestBlockRound)
+		if flushIntervalExceed || loadCompleted {
+			// adjust the last flush time, so that we would not hold off the flushing due to "working too fast"
+			tr.driver.lastFlushTime = time.Now().Add(-balancesFlushInterval)
+
+			if !rollbackSynchronousMode {
+				// switch to rebuild synchronous mode to improve performance
+				tr.dbs.Wdb.SetSynchronousMode(context.Background(), tr.accountsRebuildSynchronousMode, tr.accountsRebuildSynchronousMode >= db.SynchronousModeFull)
+
+				// flip the switch to rollback the synchronous mode once we're done.
+				rollbackSynchronousMode = true
+			}
+
+			var roundsBehind basics.Round
+
+			// flush the account data
+			tr.scheduleCommit(blk.Round(), basics.Round(config.Consensus[blk.BlockHeader.CurrentProtocol].MaxBalLookback))
+			// wait for the writing to complete.
+			tr.waitAccountsWriting()
+
+			func() {
+				tr.mu.RLock()
+				defer tr.mu.RUnlock()
+
+				// The au.dbRound after writing should be ~320 behind the block round.
+				roundsBehind = blk.Round() - tr.dbRound
+			}()
+
+			// are we too far behind ? ( taking into consideration the catchpoint writing, which can stall the writing for quite a bit )
+			if roundsBehind > initializeCachesRoundFlushInterval+basics.Round(tr.driver.catchpointInterval) {
+				// we're unable to persist changes. This is unexpected, but there is no point in keep trying batching additional changes since any further changes
+				// would just accumulate in memory.
+				close(blockEvalFailed)
+				tr.log.Errorf("initializeCaches was unable to fill up the account caches accounts round = %d, block round = %d. See above error for more details.", blk.Round()-roundsBehind, blk.Round())
+				err = fmt.Errorf("initializeCaches failed to initialize the account data caches")
+				return
+			}
+
+			// and once we flushed it to disk, update the lastFlushedRound
+			lastFlushedRound = blk.Round()
+		}
+
+		// if enough time have passed since the last time we wrote a message to the log file then give the user an update about the progess.
+		if time.Now().Sub(lastProgressMessage) > accountsCacheLoadingMessageInterval {
+			// drop the initial message if we're got to this point since a message saying "still initializing" that comes after "is initializing" doesn't seems to be right.
+			select {
+			case skipAccountCacheMessage <- struct{}{}:
+				// if we got to this point, we should be able to close the writeAccountCacheMessageCompleted channel to have the "completed initializing" message written.
+				close(writeAccountCacheMessageCompleted)
+			default:
+			}
+			tr.log.Infof("initializeCaches is still initializing account data caches, %d rounds loaded out of %d rounds", blk.Round()-lastBalancesRound, lastestBlockRound-lastBalancesRound)
+			lastProgressMessage = time.Now()
+		}
+
+		// prepare for the next iteration.
+		accLedgerEval.prevHeader = *delta.Hdr
+	}
+
+	if blockRetrievalError != nil {
+		err = blockRetrievalError
+	}
+	return
+
 }
 
 func (tr *trackerRegistry) newBlock(blk bookkeeping.Block, delta ledgercore.StateDelta) {
@@ -221,26 +395,39 @@ func (tr *trackerRegistry) newBlock(blk bookkeeping.Block, delta ledgercore.Stat
 
 func (tr *trackerRegistry) committedUpTo(rnd basics.Round) basics.Round {
 	minBlock := rnd
-
+	maxLookback := basics.Round(0)
 	for _, lt := range tr.trackers {
-		retain := lt.committedUpTo(rnd)
-		if retain < minBlock {
-			minBlock = retain
+		retainRound, lookback := lt.committedUpTo(rnd)
+		if retainRound < minBlock {
+			minBlock = retainRound
+		}
+		if lookback > maxLookback {
+			maxLookback = lookback
 		}
 	}
+
+	tr.scheduleCommit(rnd, maxLookback)
 
 	return minBlock
 }
 
-func (tr *trackerRegistry) scheduleCommit(blockqRound basics.Round) {
+func (tr *trackerRegistry) scheduleCommit(blockqRound, maxLookback basics.Round) {
 	tr.mu.RLock()
 	dbRound := tr.dbRound
 	tr.mu.RUnlock()
 
-	dc := tr.driver.produceCommittingTask(blockqRound, dbRound)
-	if dc.offset != 0 {
+	dcc := &deferredCommitContext{
+		lookback: maxLookback,
+	}
+	for _, lt := range tr.trackers {
+		dcc = lt.produceCommittingTask(blockqRound, dbRound, dcc)
+		if dcc == nil {
+			break
+		}
+	}
+	if dcc != nil {
 		tr.accountsWriting.Add(1)
-		tr.committedOffset <- dc
+		tr.deferredCommits <- dcc
 	}
 }
 
@@ -270,15 +457,15 @@ func (tr *trackerRegistry) close() {
 
 // commitSyncer is the syncer go-routine function which perform the database updates. Internally, it dequeues deferredCommits and
 // send the tasks to commitRound for completing the operation.
-func (tr *trackerRegistry) commitSyncer(deferredCommits chan deferredCommit) {
+func (tr *trackerRegistry) commitSyncer(deferredCommits chan *deferredCommitContext) {
 	defer close(tr.commitSyncerClosed)
 	for {
 		select {
-		case committedOffset, ok := <-deferredCommits:
+		case commit, ok := <-deferredCommits:
 			if !ok {
 				return
 			}
-			tr.commitRound(committedOffset)
+			tr.commitRound(commit)
 		case <-tr.ctx.Done():
 			// drain the pending commits queue:
 			drained := false
@@ -295,14 +482,13 @@ func (tr *trackerRegistry) commitSyncer(deferredCommits chan deferredCommit) {
 	}
 }
 
-func (tr *trackerRegistry) commitRound(dc deferredCommit) {
+func (tr *trackerRegistry) commitRound(dcc *deferredCommitContext) {
 	defer tr.accountsWriting.Done()
-
 	tr.mu.RLock()
 
-	offset := dc.offset
-	dbRound := dc.dbRound
-	lookback := dc.lookback
+	offset := dcc.offset
+	dbRound := dcc.oldBase
+	lookback := dcc.lookback
 
 	// we can exit right away, as this is the result of mis-ordered call to committedUpTo.
 	if tr.dbRound < dbRound || offset < uint64(tr.dbRound-dbRound) {
@@ -328,16 +514,13 @@ func (tr *trackerRegistry) commitRound(dc deferredCommit) {
 	dbRound = tr.dbRound
 	newBase := basics.Round(offset) + dbRound
 
-	dcc := deferredCommitContext{
-		offset:    offset,
-		oldBase:   dbRound,
-		newBase:   newBase,
-		lookback:  lookback,
-		flushTime: time.Now(),
-	}
+	dcc.offset = offset
+	dcc.oldBase = dbRound
+	dcc.newBase = newBase
+	dcc.flushTime = time.Now()
 
 	for _, lt := range tr.trackers {
-		err := lt.prepareCommit(&dcc)
+		err := lt.prepareCommit(dcc)
 		if err != nil {
 			tr.log.Errorf(err.Error())
 			tr.mu.RUnlock()
@@ -350,7 +533,7 @@ func (tr *trackerRegistry) commitRound(dc deferredCommit) {
 	ledgerCommitroundCount.Inc(nil)
 	err := tr.dbs.Wdb.Atomic(func(ctx context.Context, tx *sql.Tx) (err error) {
 		for _, lt := range tr.trackers {
-			err0 := lt.commitRound(ctx, tx, &dcc)
+			err0 := lt.commitRound(ctx, tx, dcc)
 			if err0 != nil {
 				return err0
 			}
@@ -373,7 +556,7 @@ func (tr *trackerRegistry) commitRound(dc deferredCommit) {
 	tr.mu.Lock()
 	tr.dbRound = newBase
 	for _, lt := range tr.trackers {
-		lt.postCommit(dcc)
+		lt.postCommit(tr.ctx, dcc)
 	}
 	tr.mu.Unlock()
 }
